+++ pyengine/Components/LifeComponent.py
@@ -4,10 +4,11 @@ __all__ = ["LifeComponent"]
 
 
 class LifeComponent:
-    def __init__(self, maxlife: int = 100):
+    def __init__(self, maxlife: int = 100, callback=None):
         self.entity = None
         self.maxlife = maxlife
         self.life = maxlife
+        self.callback = callback
 
     @property
     def entity(self):
@@ -24,6 +25,8 @@ class LifeComponent:
     @life.setter
     def life(self, life):
         self.__life = clamp(life, 0, self.maxlife)
+        if self.__life == 0 and self.callback is not None:
+            self.callback()
 
     @property
     def maxlife(self):